requirements.txt
# Core dependencies
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.19.0

# Database
psycopg[binary]>=3.1.0

# Web crawling and JavaScript rendering
aiohttp>=3.8.0
playwright>=1.40.0

# AI/API clients
openai>=1.0.0
anthropic>=0.7.0
perplexity>=0.0.0

# Utilities
lxml>=4.9.0
html5lib>=1.1
